Finalizacion de rol administrador con todos los cambios echos

+++ Proyecto_laravell/app/Http/Controllers/UserController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\RestaurarPassword;
 use App\Models\Pais;
 use App\Models\User;
 use App\Models\Candidato;
@@ -12,10 +11,13 @@ use App\Models\Reclutador;
 use App\Models\Departamento;
 use App\Models\SuperUsuario;
 use Illuminate\Http\Request;
+use App\Models\Seleccionador;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use App\Http\Requests\RestaurarPassword;
 
-class UserController extends Controller{
+class UserController extends Controller
+{
     // -------------------------- METODOS CREATE --------------------------
     // Este metodo es el que retorna la vista para registrar un candidato, 
     // Se hace el llamado de los modelos de pais, departamento y municipio
@@ -45,7 +47,8 @@ class UserController extends Controller{
     // eliminar alguno de estos se debe de acceder a este metodo con su id
     // correspondiente para buscar dicho registro por medio del metodo estatico
     // find
-    public function destroy($id){
+    public function destroy($id)
+    {
         $user = User::find($id);
         $user->delete();
 
@@ -62,10 +65,13 @@ class UserController extends Controller{
     // es el caso se elimina dicho registro, para luego validar cual fue el
     // rol al cual cambio y al registro que queda en la tabla usuario se actualiza
     // el campo role_id para luego salvarlo en su tabla correspondiente 
-    public function cambioRol(Request $request, $id){
+    public function cambioRol(Request $request, $id)
+    {
         $usuario = User::find($id);
-
         if ($usuario->role_id == 1) {
+            if($request->menu == 1){
+                return redirect()->back();
+            }
             $administrador = SuperUsuario::where('user_id',  $usuario->id)->first();
             $administrador->delete();
 
@@ -87,18 +93,16 @@ class UserController extends Controller{
 
                 $instructor->save();
                 return redirect()->back();
-            }
-            // else if($request -> menu == 4){
-            //     $usuario -> role_id = $request -> menu;
-            //     $usuario -> save();
+            } else if ($request->menu == 4) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
 
-            //     $seleccionador = new Seleccionador();
-            //     $seleccionador -> user_id = $usuario -> id;
+                $seleccionador = new Seleccionador();
+                $seleccionador->user_id = $usuario->id;
 
-            //     $seleccionador->save();
-            //     return redirect()->route('super.index');
-            // }
-            else if ($request->menu == 5) {
+                $seleccionador->save();
+                return redirect()->back();
+            } else if ($request->menu == 5) {
                 $usuario->role_id = $request->menu;
                 $usuario->save();
 
@@ -113,6 +117,9 @@ class UserController extends Controller{
         // Validacion donde se comprueba si el usuario al que se desea 
         // cambiar el rol es un candidato
         if ($usuario->role_id == 2) {
+            if($request->menu == 2){
+                return redirect()->back();
+            }
             $candidato = Candidato::where('user_id',  $usuario->id)->first();
             $candidato->delete();
 
@@ -134,29 +141,33 @@ class UserController extends Controller{
 
                 $instructor->save();
                 return redirect()->back();
+            } else if ($request->menu == 4) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $seleccionador = new Seleccionador();
+                $seleccionador->user_id = $usuario->id;
+
+                $seleccionador->save();
+                return redirect()->back();
+            } else if ($request->menu == 5) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $reclutador = new Reclutador();
+                $reclutador->user_id = $usuario->id;
+
+                $reclutador->save();
+                return redirect()->back();
             }
-            // else if($request -> menu == 4){
-            //     $usuario -> role_id = $request -> menu;
-            //     $usuario -> save();
-
-            //     $seleccionador = new Seleccionador();
-            //     $seleccionador -> user_id = $usuario -> id;
-
-            //     $seleccionador->save();
-            //     return redirect()->route('super.index');
-            // }
-            // else if($request -> menu == 5){
-            //     $usuario -> role_id = $request -> menu;
-            //     $usuario -> save();
-
-            $reclutador = new Reclutador();
-            $reclutador->user_id = $usuario->id;
-
-            $reclutador->save();
-            return redirect()->route('super.index');
         }
 
+        // Validacion donde se comprueba si el usuario al que se desea 
+        // cambiar el rol es un Instructor
         if ($usuario->role_id == 3) {
+            if($request->menu == 3){
+                return redirect()->back();
+            }
             $instructor = Instructor::where('user_id',  $usuario->id)->first();
             $instructor->delete();
 
@@ -179,18 +190,16 @@ class UserController extends Controller{
 
                 $candidato->save();
                 return redirect()->back();
-            }
-            // else if($request -> menu == 4){
-            //     $usuario -> role_id = $request -> menu;
-            //     $usuario -> save();
+            } else if ($request->menu == 4) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
 
-            //     $seleccionador = new Seleccionador();
-            //     $seleccionador -> user_id = $usuario -> id;
+                $seleccionador = new Seleccionador();
+                $seleccionador->user_id = $usuario->id;
 
-            //     $seleccionador->save();
-            //     return redirect()->route('super.index');
-            // }
-            else if ($request->menu == 5) {
+                $seleccionador->save();
+                return redirect()->back();
+            } else if ($request->menu == 5) {
                 $usuario->role_id = $request->menu;
                 $usuario->save();
 
@@ -198,15 +207,20 @@ class UserController extends Controller{
                 $reclutador->user_id = $usuario->id;
 
                 $reclutador->save();
-                return redirect()->route('super.index');
+                return redirect()->back();
             }
         }
 
-        if ($usuario->role_id == 5){
-            $reclutador = Reclutador::where('user_id', $usuario->id)->first();
+        // Validacion donde se comprueba si el usuario al que se desea 
+        // cambiar el rol es un seleccionador
+        if ($usuario->role_id == 4) {
+            if($request->menu == 4){
+                return redirect()->back();
+            }
+            $reclutador = Seleccionador::where('user_id', $usuario->id)->first();
             $reclutador->delete();
 
-            if($request->menu == 1){
+            if ($request->menu == 1) {
                 $usuario->role_id = $request->menu;
                 $usuario->save();
 
@@ -214,9 +228,82 @@ class UserController extends Controller{
                 $administrador->user_id = $usuario->id;
                 $administrador->save();
                 return redirect()->back();
+            } else if ($request->menu == 3) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $instructor = new Instructor();
+                $instructor->user_id = $usuario->id;
+
+                $instructor->save();
+                return redirect()->back();
+            } else if ($request->menu == 2) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $candidato = new Candidato();
+                $candidato->user_id = $usuario->id;
+
+                $candidato->save();
+                return redirect()->back();
+            } else if ($request->menu == 5) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $reclutador = new Reclutador();
+                $reclutador->user_id = $usuario->id;
+
+                $reclutador->save();
+                return redirect()->back();
+            }
+        }
+
+        if ($usuario->role_id == 5) {
+            if($request->menu == 5){
+                return redirect()->back();
+            }
+            $reclutador = Reclutador::where('user_id', $usuario->id)->first();
+            $reclutador->delete();
+
+            if ($request->menu == 1) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $administrador = new SuperUsuario();
+                $administrador->user_id = $usuario->id;
+                $administrador->save();
+                return redirect()->back();
+            } else if ($request->menu == 3) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $instructor = new Instructor();
+                $instructor->user_id = $usuario->id;
+
+                $instructor->save();
+                return redirect()->back();
+            } else if ($request->menu == 2) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $candidato = new Candidato();
+                $candidato->user_id = $usuario->id;
+
+                $candidato->save();
+                return redirect()->back();
+            } else if ($request->menu == 4) {
+                $usuario->role_id = $request->menu;
+                $usuario->save();
+
+                $seleccionador = new Seleccionador();
+                $seleccionador->user_id = $usuario->id;
+
+                $seleccionador->save();
+                return redirect()->back();
             }
         }
     }
+
     // -------------------------------------------------------------------
 
 
@@ -224,7 +311,8 @@ class UserController extends Controller{
     // Este metodo retornara una vista que pedira la contraseña actual que
     // tenga el usuario, la nueva contraseña y la confirmacion de la nueva
     // contraseña
-    public function restaurarCreate(){
+    public function restaurarCreate()
+    {
         return view('auth.restaurar');
     }
     // -------------------------------------------------------------------
@@ -241,14 +329,15 @@ class UserController extends Controller{
     // nueva es igual a la contraseña de verificacion para luego asi proceder
     // a cambiar la contraseña del usuario autenticado hasheandola, para asi
     // cerrarle la sesion y volverlo a la ruta login
-    public function restaurarContraseña(RestaurarPassword $request){
+    public function restaurarContraseña(RestaurarPassword $request)
+    {
         $usuario = Auth::user();
 
         // Hash::check hace la validacion entre los dos parametros que
         // se le asignen y deshashea las contraseñas que esten hashadas
         // para asi poder realizar la validacion
-        if (Hash::check($request->contraseña_old, $usuario->password)){
-            if ($request->contraseña_new == $request->contraseña_verifi){
+        if (Hash::check($request->contraseña_old, $usuario->password)) {
+            if ($request->contraseña_new == $request->contraseña_verifi) {
 
                 $usuarioNewPassword = User::find($usuario->id);
                 $usuarioNewPassword->password = Hash::make($request->contraseña_new);
@@ -258,7 +347,7 @@ class UserController extends Controller{
                 auth()->logout();
                 session()->invalidate();
                 return redirect()->route('login');
-            }else{
+            } else {
                 return redirect()->back();
             }
         };
@@ -273,16 +362,17 @@ class UserController extends Controller{
     // de esta manera para no hacer el mismo codigo en cada uno de los roles
     // si no que en un solo metodo validar el usuario que este auntenticado
     // y dependiendo de su rol le retorne la vista de inicio que le corresponda
-    public function restaurarHome(){
+    public function restaurarHome()
+    {
         $user = Auth::user();
 
         if ($user->role_id == 1) {
             return redirect()->route('super.index');
-        }elseif($user->role_id == 2) {
+        } elseif ($user->role_id == 2) {
             return redirect()->route('candidato.index');
-        }elseif($user->role_id == 4){
+        } elseif ($user->role_id == 4) {
             return redirect()->route('seleccionador.index');
-        }elseif($user->role_id == 5) {
+        } elseif ($user->role_id == 5) {
             return redirect()->route('reclutador.index');
         }
     }
